LightcullingPass for forward+ is now separate from Renderer.cpp.
