+++ include/Engine/Rendering/Renderer.h
@@ -12,11 +12,7 @@
 #include "../Core/World.h"
 #include "PickingPass.h"
 #include "DrawScenePass.h"
-
-
-#define TILE_SIZE 16 
-#define NUM_LIGHTS 1000
-
+#include "LightCullingPass.h"
 
 #include "../Core/EventBroker.h"
 #include "EPicking.h"
@@ -46,12 +42,12 @@ private:
 
     DrawScenePass* m_DrawScenePass;
     PickingPass* m_PickingPass;
+    LightCullingPass* m_LightCullingPass;
 
     //----------------------Functions----------------------//
     void InitializeWindow();
     void InitializeShaders();
     void InitializeTextures();
-    void InitializeSSBOs();
     void InitializeRenderPasses();
     //TODO: Renderer: Get InputUpdate out of renderer
     void InputUpdate(double dt);
@@ -59,58 +55,18 @@ private:
     void DrawScreenQuad(GLuint textureToDraw);
 
     //----------------------Forward+-----------------------//
-    void CalculateFrustum();
-    void CullLights();
     void DrawForwardPlus(RenderQueueCollection& rq);
     //Frustum
-    struct Plane {
-        glm::vec3 Normal;
-        float d;
-    };
-
-    struct Frustum {
-        Plane Planes[4];
-    };
-    Frustum m_Frustums[80*45]; //TODO: Renderer: Make this change with resolution
-
-    //Lights
-    void TEMPCreateLights();
-    //TODO: Renderer: Add Directionllights, spotlights and area lights to this as type.
-    struct PointLight {
-        glm::vec4 Position = glm::vec4(0.f);
-        glm::vec4 Color = glm::vec4(1.f);
-        float Radius = 5.f;
-        float Intensity = 0.8f;
-        float Falloff = 0.3f;
-        float Padding = 1337;
-    };
-    PointLight m_PointLights[NUM_LIGHTS];
-
-    struct LightGrid {
-        float Start;
-        float Amount;
-        glm::vec2 Padding;
-    };
-
-    LightGrid m_LightGrid[80*45];
-
-    int m_LightOffset = 0;
-
-    float m_LightIndex[80*45*200];
+    
 
     //-------------------------SSBO------------------------//
-    GLuint m_FrustumSSBO = 0;
-    GLuint m_LightSSBO = 0;
-    GLuint m_LightGridSSBO = 0;
-    GLuint m_LightOffsetSSBO = 0;
-    GLuint m_LightIndexSSBO = 0;
+
 
     void GenerateTexture(GLuint* texture, GLenum wrapping, GLenum filtering, glm::vec2 dimensions, GLint internalFormat, GLint format, GLenum type);
 	//--------------------ShaderPrograms-------------------//
 	ShaderProgram* m_BasicForwardProgram;
     ShaderProgram* m_DrawScreenQuadProgram;
-    ShaderProgram* m_CalculateFrustumProgram;
-    ShaderProgram* m_LightCullProgram;
+
     ShaderProgram* m_ForwardPlusProgram;
 
 };
